refactor: Apply GitHub mixins pattern to BitBucket service (#10728)
