+++ openhands/integrations/bitbucket/service/__init__.py
@@ -1,0 +1,13 @@
+from .base import BitBucketMixinBase
+from .branches import BitBucketBranchesMixin
+from .features import BitBucketFeaturesMixin
+from .prs import BitBucketPRsMixin
+from .repos import BitBucketReposMixin
+
+__all__ = [
+    'BitBucketMixinBase',
+    'BitBucketBranchesMixin',
+    'BitBucketFeaturesMixin',
+    'BitBucketPRsMixin',
+    'BitBucketReposMixin',
+]
